+++ strings/pascal_case.go
@@ -1,0 +1,8 @@
+package strings
+
+// Converts string to pascal case.
+//
+// Complexity: O(n)
+func PascalCase(input string) string {
+	return internalCamelCase(input, true)
+}